refactor(base): split outbox/tests.rs by behaviour (#244)
outbox/tests.rs was 734 lines (over the 700-line cap). Split by what
each group exercises:

  tests/mod.rs           - file-level docs + sub-module declarations
  tests/payload.rs       - CellOutboxPayload serde + event_type contract (7 tests)
  tests/row_to_message.rs- row -> BaseToCellMsg decoder (5 tests)
  tests/live_db.rs       - end-to-end SQL paths gated on DATABASE_URL (6 tests)

The live-DB sentinel-base, cleanup helpers, and `drain_item_used_for`
are kept inside `live_db.rs` because they're only consumed by those
tests. The pure-logic groups don't need any shared fixtures.

All 18 tests pass; no behavior change.
